Started refactor of column storage
