+++ src/data/sqlite/statements/columns.createTable.js
